Consolidating graphUtil.js and Shared.js
